requirements.txt
mysql
requests
lxml